% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSPoTApi.r
\name{VariantSPoTApi_add_atomic_variant}
\alias{VariantSPoTApi_add_atomic_variant}
\title{Create a single new object from a given data file with optional supplied metadata.}
\usage{
VariantSPoTApi_add_atomic_variant(source, ...)
}
\description{
Create a single new object from a given data file with optional supplied metadata.
}
