% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSPoTApi.r
\name{VariantSPoTApi_add_variants}
\alias{VariantSPoTApi_add_variants}
\title{Create multiple new objects from a multi-row data file with optional supplied metadata}
\usage{
VariantSPoTApi_add_variants(source, ...)
}
\arguments{
\item{source}{}
}
\description{
Create multiple new objects from a multi-row data file with optional supplied metadata
}
