% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSPoTApi.r
\name{VariantSPoTApi_delete_atomic}
\alias{VariantSPoTApi_delete_atomic}
\title{Delete the object}
\usage{
VariantSPoTApi_delete_atomic(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Delete the object
}
