% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSPoTApi.r
\name{VariantSPoTApi_update_variant}
\alias{VariantSPoTApi_update_variant}
\title{Update object metadata}
\usage{
VariantSPoTApi_update_variant(id, body, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}

\item{body}{Metadata in the form of \code{{key: value, key2: value2, ...}}}
}
\description{
Update object metadata
}
